Switch to PSR-4.
