+++ tests/AbstractTestCase.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Tests;
+
+use Mockery;
+use PHPUnit_Framework_TestCase;
+
+abstract class AbstractTestCase extends PHPUnit_Framework_TestCase
+{
+    public function tearDown()
+    {
+        Mockery::close();
+
+        parent::tearDown();
+    }
+}
