+++ tests/BuilderTest.php
@@ -1,19 +1,18 @@
 <?php
 
+namespace Tests;
+
+use Mockery;
+use StdClass;
+use Laravel\Scout\Builder;
+use Illuminate\Pagination\Paginator;
 use Illuminate\Database\Eloquent\Collection;
 
-use Illuminate\Pagination\Paginator;
-
-class BuilderTest extends PHPUnit_Framework_TestCase
+class BuilderTest extends AbstractTestCase
 {
-    public function tearDown()
-    {
-        Mockery::close();
-    }
-
     public function test_pagination_correctly_handles_paginated_results()
     {
-        $builder = new Laravel\Scout\Builder($model = Mockery::mock(), 'zonda');
+        $builder = new Builder($model = Mockery::mock(), 'zonda');
         $model->shouldReceive('searchableUsing')->andReturn($engine = Mockery::mock());
         Paginator::currentPageResolver(function () {
             return 1;
@@ -24,6 +23,6 @@ class BuilderTest extends PHPUnit_Framework_TestCase
         $engine->shouldReceive('paginate');
         $engine->shouldReceive('map')->andReturn(Collection::make([new StdClass]));
 
-        $paginator = $builder->paginate();
+        $builder->paginate();
     }
 }
