+++ tests/MakeSearchableTest.php
@@ -1,18 +1,17 @@
 <?php
 
+namespace Tests;
+
+use Mockery;
+use Laravel\Scout\Jobs\MakeSearchable;
 use Illuminate\Database\Eloquent\Collection;
 
-class MakeSearchableTest extends PHPUnit_Framework_TestCase
+class MakeSearchableTest extends AbstractTestCase
 {
-    public function tearDown()
-    {
-        Mockery::close();
-    }
-
     public function test_handle_passes_the_collection_to_engine()
     {
-        $job = new Laravel\Scout\Jobs\MakeSearchable($collection = Collection::make([
-            $model = Mockery::mock()
+        $job = new MakeSearchable($collection = Collection::make([
+            $model = Mockery::mock(),
         ]));
 
         $model->shouldReceive('searchableUsing->update')->with($collection);
