+++ tests/ModelObserverTest.php
@@ -1,19 +1,15 @@
 <?php
 
-use Illuminate\Database\Eloquent\Collection;
+namespace Tests;
 
-use Illuminate\Pagination\Paginator;
+use Mockery;
+use Laravel\Scout\ModelObserver;
 
-class ModelObserverTest extends PHPUnit_Framework_TestCase
+class ModelObserverTest extends AbstractTestCase
 {
-    public function tearDown()
-    {
-        Mockery::close();
-    }
-
     public function test_created_handler_makes_model_searchable()
     {
-        $observer = new Laravel\Scout\ModelObserver;
+        $observer = new ModelObserver;
         $model = Mockery::mock();
         $model->shouldReceive('searchable');
         $observer->created($model);
@@ -21,7 +17,7 @@ class ModelObserverTest extends PHPUnit_Framework_TestCase
 
     public function test_created_handler_doesnt_make_model_searchable_when_disabled()
     {
-        $observer = new Laravel\Scout\ModelObserver;
+        $observer = new ModelObserver;
         $model = Mockery::mock();
         $observer->disableSyncingFor(get_class($model));
         $model->shouldReceive('searchable')->never();
@@ -30,7 +26,7 @@ class ModelObserverTest extends PHPUnit_Framework_TestCase
 
     public function test_updated_handler_makes_model_searchable()
     {
-        $observer = new Laravel\Scout\ModelObserver;
+        $observer = new ModelObserver;
         $model = Mockery::mock();
         $model->shouldReceive('searchable');
         $observer->updated($model);
@@ -38,7 +34,7 @@ class ModelObserverTest extends PHPUnit_Framework_TestCase
 
     public function test_deleted_handler_makes_model_unsearchable()
     {
-        $observer = new Laravel\Scout\ModelObserver;
+        $observer = new ModelObserver;
         $model = Mockery::mock();
         $model->shouldReceive('unsearchable');
         $observer->deleted($model);
@@ -46,7 +42,7 @@ class ModelObserverTest extends PHPUnit_Framework_TestCase
 
     public function test_restored_handler_makes_model_searchable()
     {
-        $observer = new Laravel\Scout\ModelObserver;
+        $observer = new ModelObserver;
         $model = Mockery::mock();
         $model->shouldReceive('searchable');
         $observer->restored($model);
